+++ tasty/runtime/data/parameter/Enum.scala
@@ -6,23 +6,23 @@ trait Enum {
   }
 
   enum TypeParameters[A, B] {
-    case Case extends TypeParameters[Int, Int]
+    case Case extends TypeParameters[Int, Long]
   }
 
   enum ValueParameter(x: Int) {
     case Case extends ValueParameter/**/(1)/**/
   }
 
-  enum ValueParameters(x: Int, y: Int) {
-    case Case extends ValueParameters/**/(1, 2)/**/
+  enum ValueParameters(x: Int, y: Long) {
+    case Case extends ValueParameters/**/(1, 2L)/**/
   }
 
   enum EmptyClause/**/()/**/ {
     case Case extends EmptyClause
   }
 
-  enum MultipleClauses(x: Int)(y: Int) {
-    case Case extends MultipleClauses/**/(1)(2)/**/
+  enum MultipleClauses(x: Int)(y: Long) {
+    case Case extends MultipleClauses/**/(1)(2L)/**/
   }
 
   enum TypeAndValueParameters[A](x: Int) {
